Move lxdclient package under LXD provider.
