+++ provider/lxd/environ_raw.go
@@ -8,8 +8,8 @@ package lxd
 import (
 	"github.com/juju/errors"
 
-	"github.com/juju/juju/container/lxd/lxdclient"
 	"github.com/juju/juju/provider/common"
+	"github.com/juju/juju/provider/lxd/lxdclient"
 )
 
 type rawProvider struct {
