+++ provider/lxd/export_test.go
@@ -6,8 +6,8 @@
 package lxd
 
 import (
-	"github.com/juju/juju/container/lxd/lxdclient"
 	"github.com/juju/juju/environs"
+	"github.com/juju/juju/provider/lxd/lxdclient"
 )
 
 var (
