+++ provider/lxd/instance_test.go
@@ -10,9 +10,9 @@ import (
 	jc "github.com/juju/testing/checkers"
 	gc "gopkg.in/check.v1"
 
-	"github.com/juju/juju/container/lxd/lxdclient"
 	"github.com/juju/juju/instance"
 	"github.com/juju/juju/provider/lxd"
+	"github.com/juju/juju/provider/lxd/lxdclient"
 )
 
 type instanceSuite struct {
