+++ provider/lxd/lxdclient/addserver.go
