+++ provider/lxd/lxdclient/addserver_test.go
