+++ provider/lxd/lxdclient/lxd_client.go
