+++ provider/lxd/lxdclient/utils.go
